# portal/public/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Omni Runner Portal API",
    "description": "API do portal de assessorias esportivas Omni Runner. Gerencia atletas, treinos, financeiro, OmniCoins e administração da plataforma.",
    "version": "1.0.0",
    "contact": {
      "name": "Omni Runner"
    }
  },
  "servers": [
    {
      "url": "https://portal.omnirunner.app",
      "description": "Produção"
    }
  ],
  "security": [
    { "bearerAuth": [] }
  ],
...
  "paths": {
...
    "/api/swap": {
      "get": {
        "tags": ["OmniCoins"],
        "summary": "Listar ofertas de swap abertas",
        "description": "Lista ofertas de swap (USD por créditos) disponíveis para a assessoria.",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": { "offers": { "type": "array", "items": { "type": "object" } } }
                }
              }
            }
          },
          "400": { "description": "Grupo não selecionado" },
          "401": { "description": "Não autorizado" },
          "403": { "description": "Apenas admin_master" },
          "429": { "description": "Rate limit excedido" }
        }
      },
      "post": {
        "tags": ["OmniCoins"],
        "summary": "Criar, aceitar ou cancelar swap",
        "description": "Cria nova oferta, aceita ou cancela oferta existente.",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "oneOf": [
                  {
                    "type": "object",
                    "required": ["action", "amount_usd"],
                    "properties": {
                      "action": { "type": "string", "enum": ["create"] },
                      "amount_usd": { "type": "number", "minimum": 100, "maximum": 500000 }
                    }
                  },
                  {
                    "type": "object",
                    "required": ["action", "order_id"],
                    "properties": {
                      "action": { "type": "string", "enum": ["accept", "cancel"] },
                      "order_id": { "type": "string", "format": "uuid" }
                    }
                  }
                ]
              }
            }
          }
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "oneOf": [
                    { "type": "object", "properties": { "order": { "type": "object" } } },
                    { "type": "object", "properties": { "ok": { "type": "boolean" } } }
                  ]
                }
              }
            }
          },
          "400": { "description": "Validação falhou" },
          "401": { "description": "Não autorizado" },
          "403": { "description": "Apenas admin_master" },
          "422": { "description": "Operação falhou" },
          "429": { "description": "Rate limit excedido" },
          "503": { "description": "Swap não disponível" }
        }
      }
    },
...
    "/api/team/remove": {
      "post": {
        "tags": ["Team"],
        "summary": "Remover membro",
        "description": "Remove membro da assessoria. Não é possível remover a si mesmo ou outro admin_master.",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["member_id"],
                "properties": { "member_id": { "type": "string" } }
              }
            }
          }
        },
        "responses": {
          "200": { "description": "Membro removido" },
          "400": { "description": "Validação falhou ou tentativa de remover a si mesmo" },
          "401": { "description": "Não autorizado" },
          "403": { "description": "Apenas admin_master ou tentativa de remover admin" },
          "404": { "description": "Membro não encontrado" },
          "429": { "description": "Rate limit excedido" }
        }
      }
    },
...
    "/api/staff-alerts": {
      "get": {
        "tags": ["CRM"],
        "summary": "Alertas da equipe",
        "description": "Retorna alertas (atletas inativos, créditos baixos, etc.).",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "alerts": {
                      "type": "array",
                      "items": {
                        "type": "object",
                        "properties": {
                          "key": { "type": "string" },
                          "label": { "type": "string" },
                          "count": { "type": "integer" },
                          "severity": { "type": "string", "enum": ["warning", "error", "info"] }
                        }
                      }
                    }
                  }
                }
              }
            }
          },
          "400": { "description": "Grupo não selecionado" },
          "401": { "description": "Não autorizado" },
          "403": { "description": "Apenas admin/coach" },
          "429": { "description": "Rate limit excedido" }
        }
      }
    },
...
    "/api/platform/support": {
      "post": {
        "tags": ["Platform Admin"],
        "summary": "Gestão de suporte",
        "description": "Responder, fechar ou reabrir ticket de suporte.",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["action", "ticket_id"],
                "properties": {
                  "action": { "type": "string", "enum": ["reply", "close", "reopen"] },
                  "ticket_id": { "type": "string" },
                  "message": { "type": "string" }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "status": { "type": "string", "enum": ["replied", "closed", "reopened"] },
                    "ticket_id": { "type": "string" }
                  }
                }
              }
            }
          },
          "400": { "description": "action/ticket_id obrigatórios ou message obrigatório para reply" },
          "401": { "description": "Não autenticado" },
          "403": { "description": "Não é admin da plataforma" },
          "429": { "description": "Rate limit excedido" }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "Token JWT do Supabase (access_token da sessão)"
      }
    },
...
  }
}